backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
tortoise-orm==0.20.0
aiosqlite==0.17.0
pydantic==2.5.2
email-validator==2.1.0.post1 